+++ tests/AvroConvertTests/Json2Avro/GenericJson2AvroTests.cs
@@ -4,7 +4,7 @@ using Newtonsoft.Json;
 using SolTechnology.Avro;
 using Xunit;
 
-namespace AvroConvertComponentTests.JsonToAvro
+namespace AvroConvertComponentTests.Json2Avro
 {
     public class GenericJson2AvroTests
     {